\name{EnsDb.Ggallus.v110}
\docType{package}

\alias{EnsDb.Ggallus.v110-package}
\alias{EnsDb.Ggallus.v110}
\alias{EnsDb.Ggallus.v110}


\title{Ensembl based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at Ensembl on
  Thu Oct 19 08:26:38 2023 and based on the 110
}

\author{Haixu Jiang}



\examples{
## load the library
##library(EnsDb.Ggallus.v110)
## list the contents that are loaded into memory
ls('package:EnsDb.Ggallus.v110')
## show the db object that is loaded by calling it's name
EnsDb.Ggallus.v110

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
